requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
jinja2>=3.1.4
pydantic-settings>=2.6.0
pyyaml>=6.0.2
itsdangerous>=2.2.0
python-multipart>=0.0.9
